add error submodules
